Move reporting to its own module
